# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clawith system OpenAPI",
    "version": "1",
    "description": "Generated from FastAPI. Context interactions, instance sessions, scene selection and optional per-message H5 host context. Check capabilities on the target environment for availability."
  },
  "paths": {
    "/api/openapi/v1/auth/token": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Token",
        "operationId": "token_api_openapi_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": [
                      "client_credentials"
                    ]
                  },
                  "scope": {
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "grant_type"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthTokenOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ClientSecretBasic": []
          },
          {
            "ClientSecretBasic": []
          }
        ]
      }
    },
    "/api/openapi/v1/auth/revoke": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Revoke Token",
        "operationId": "revoke_token_api_openapi_v1_auth_revoke_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "token_type_hint": {
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "token"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "ClientSecretBasic": []
          }
        ]
      }
    },
    "/api/openapi/v1/capabilities": {
      "get": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Capabilities",
        "operationId": "capabilities_api_openapi_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "SystemOAuth2": []
          }
        ]
      }
    },
    "/api/openapi/v1/digital-employees/search": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Employees",
        "operationId": "employees_api_openapi_v1_digital_employees_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeSearchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeePageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SystemOAuth2": [
              "employees:read"
            ]
          }
        ]
      }
    },
    "/api/openapi/v1/digital-employees/{employee_id}/scenes/search": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Employee Scenes",
        "description": "List this employee's enabled published scenes using the public H5 manifest.",
        "operationId": "employee_scenes_api_openapi_v1_digital_employees__employee_id__scenes_search_post",
        "security": [
          {
            "SystemOAuth2": [
              "employees:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeScenesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/openapi/v1/digital-employees/{employee_id}/access": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Employee",
        "description": "With only user, returns the existing employee information and access URL. Adding instance_ref, interaction, embed_origin or scene_key also requires auth:login and returns a temporary login URL. An interaction prepares a first question and optional JSON context without executing it. Opening the login URL activates one new conversation and one first question; identical request_id retries reuse that interaction. instance_ref alone resumes its current conversation. scene_key selects an enabled published scene when the login URL is opened. host_context.enabled adds per-message host snapshots; instance_ref and embed_origin are required. Enabling it does not create an automatic question.",
        "operationId": "employee_api_openapi_v1_digital_employees__employee_id__access_post",
        "security": [
          {
            "SystemOAuth2": [
              "employees:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeAccessInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeAccessOut"
                }
              }
            }
          },
          "400": {
            "description": "invalid_interaction or invalid_host_context: invalid optional launch fields."
          },
          "404": {
            "description": "scene_unavailable: the selected scene is not available for activation."
          },
          "409": {
            "description": "interaction_conflict: request_id was already used with different business content."
          },
          "410": {
            "description": "interaction_expired or interaction_unavailable: the interaction can no longer be activated."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/openapi/v1/auth/links": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Login Link",
        "operationId": "login_link_api_openapi_v1_auth_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginLinkInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginLinkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SystemOAuth2": [
              "auth:login"
            ]
          }
        ]
      }
    },
    "/api/openapi/v1/auth/link-exchange": {
      "post": {
        "tags": [
          "OpenAPI v1"
        ],
        "summary": "Exchange Link",
        "description": "Exchange a single-use login code for the delegated user's ordinary login. Codes from extended employee access recheck employee permission and activate the prepared conversation before returning its redirect URI. First-question execution starts only after the activation commits; repeated activation through newly issued codes does not submit another question. Generic auth/links codes retain their existing login and redirect behavior.",
        "operationId": "exchange_link_api_openapi_v1_auth_link_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginExchangeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginExchangeOut"
                }
              }
            }
          },
          "400": {
            "description": "invalid_interaction: the launch reference is invalid."
          },
          "404": {
            "description": "scene_unavailable: the selected scene is no longer available."
          },
          "410": {
            "description": "interaction_expired or interaction_unavailable: the prepared interaction is no longer available."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginExchangeInput": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 20,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "LoginExchangeInput"
      },
      "LoginLinkOut": {
        "properties": {
          "login_url": {
            "type": "string",
            "title": "Login Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "login_url",
          "expires_in"
        ],
        "title": "LoginLinkOut"
      },
      "LoginLinkInput": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/DelegatedUser"
          },
          "redirect_uri": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Redirect Uri"
          },
          "embed_origin": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Origin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "redirect_uri"
        ],
        "title": "LoginLinkInput"
      },
...
      "EmployeeAccessInput": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/DelegatedUser"
          },
          "instance_ref": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance Ref"
          },
          "interaction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextInteractionInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "embed_origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Origin"
          },
          "scene_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Key"
          },
          "host_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostContextInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user"
        ],
        "title": "EmployeeAccessInput"
      },
...
      "EmployeeScenesOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmployeeSceneOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "EmployeeScenesOut"
      },
...
      "EmployeeUserInput": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/DelegatedUser"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user"
        ],
        "title": "EmployeeUserInput"
      },
      "EmployeePageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmployeeOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_more"
        ],
        "title": "EmployeePageOut"
      },
...
      "EmployeeSearchInput": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/DelegatedUser"
          },
          "search": {
            "type": "string",
            "maxLength": 100,
            "title": "Search",
            "default": ""
          },
          "page": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 20
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user"
        ],
        "title": "EmployeeSearchInput"
      },
      "CapabilitiesOut": {
        "properties": {
          "protocol_version": {
            "type": "integer",
            "const": 1,
            "title": "Protocol Version"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "trust_user_identity": {
            "type": "boolean",
            "title": "Trust User Identity"
          },
          "h5_launcher": {
            "$ref": "#/components/schemas/H5LauncherCapability"
          }
        },
        "type": "object",
        "required": [
          "protocol_version",
          "scopes",
          "trust_user_identity"
        ],
        "title": "CapabilitiesOut"
      },
      "H5LauncherCapability": {
        "properties": {
          "interaction": {
            "$ref": "#/components/schemas/InteractionCapability"
          },
          "instance_ref": {
            "type": "boolean",
            "const": true,
            "title": "Instance Ref",
            "default": true
          },
          "scenes": {
            "$ref": "#/components/schemas/SceneCapability"
          },
          "host_context": {
            "$ref": "#/components/schemas/InteractionCapability"
          }
        },
        "type": "object",
        "title": "H5LauncherCapability"
      },
...
      "OAuthTokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "title": "Token Type"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "scope"
        ],
        "title": "OAuthTokenOut"
      },
...
    },
    "securitySchemes": {
      "ClientSecretBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "SystemOAuth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "employees:read": "Read delegated user's visible digital employees",
              "auth:login": "Create temporary login links for delegated users"
            },
            "tokenUrl": "/api/openapi/v1/auth/token"
          }
        }
      }
    }
  }
}